chore: Update npm dependencies for Redux Logger
This commit updates the npm dependencies for Redux Logger to the latest stable version. It includes updating the package-lock.json and package.json files to include the new version of Redux Logger and its associated dependencies. This update ensures that the application is using the most up-to-date features and bug fixes provided by Redux Logger.